Embed bitcode marker in iOS binaries
This allows bitcode-enabled apps to link with FridaGadget.dylib. We only
enable markers, and not actual bitcode, as our dylib should never be
deployed to production. The reason is that we depend on the app being
launched with lldb in order to give us the runtime-equivalent of the
"run-unsigned-code" entitlement.