Fix fragile OpenSSL CPU feature detection on ARM
Instead of temporarily installing a signal handler and relying on
illegal instructions to detect what the CPU supports, only do that
as a last resort. Meaning:

- Linux: Use BoringSSL's elaborate feature detection strategy.
- i/macOS: Evolve special-case to also cover 64-bit ARM to avoid messing
  with signal handler state, which is dangerous on a live process.
- Any other OS: Use the old SIGILL-based logic.

The end result is that we fix a long-standing regression preventing
Frida from being used with libhoudini on Android, live injection is
safer on iOS, and debugging experience is better as we avoid confusing
SIGILL stops at load time.